% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lohman_conv.R
\name{dir2river}
\alias{dir2river}
\title{Conver flow direction ascii grid to river table.}
\usage{
dir2river(dir.file, grid.loc, arcinfo = TRUE)
}
\arguments{
\item{dir.file}{Path of the flow direction ascii grid file.}

\item{grid.loc}{A two-column array, a column is longitute
and another column is latitude of the grids
to be considered in XAJ model.}

\item{arcinfo}{If the direction code is arcgis style. Default
TRUE.}
}
\value{
A table containing the informations of the river
        network. Each row represents a river channel, the first column
        is the next river channel to flow in, the second column is the
        length [m] of the river channel, while the third column is the
        subbasin flow into that river channel.
}
\description{
Conver flow direction ascii grid to rivers.
}

