% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lohman_conv.R
\name{Lohmann_conv}
\alias{Lohmann_conv}
\title{Convolution of the distributed Xinanjiang model output by the routing
model of Lohman et al.}
\usage{
Lohmann_conv(Qs, uh)
}
\arguments{
\item{Qs}{Runoff time series of each subbasins.}

\item{uh}{Unit hydrographs of Lohman model of each subbasins.}
}
\value{
The routing result of the XAJ output, that is, the time series
        of the streamflow at the outlet of the river.
}
\description{
Convolution of the distributed Xinanjiang model output by
             the routing model of Lohman et al.
}

