% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XAJ.R
\name{name_params}
\alias{name_params}
\title{Set the names of the parameters.}
\usage{
name_params(params)
}
\arguments{
\item{params}{The vector of the XAJ parameters}
}
\description{
Set the names of the parameters to output so that to
             see what parameter the parameters are.
}

