% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XAJ.R
\docType{package}
\name{XAJ}
\alias{XAJ}
\alias{XAJ-package}
\title{XAJ: An R implementation of three-source Xinanjiang model by Renjun Zhao.}
\usage{
XAJ(PREC, EVAP, params, area, dt = 24, full.UH = FALSE)
}
\arguments{
\item{PREC}{Time series of precipitation (daily)}

\item{EVAP}{Time series of pan evaporation
or potential evapotranspiration (daily), length must
equal to \code{PREC}}

\item{params}{Parameters (see below)}

\item{area}{Basin area.}

\item{full.UH}{Use the unit hydrograph defined by user, rather than the
instantaneous unit hydrograph (IUH) of Nash, for routing
of surface runoff. Default FALSE.}
}
\value{
This function returns a data frame of some common variables of the XAJ
        model at each time step, such as evaporation, soil moisture, surface
        and underground runoff.

        The variables of the table including:

        E,   Total evaporation (mm)

        EU,  Evaporation (mm) of upper soil layer

        EL,  Evaporation (mm) of lower soil layer

        ED,  Evaporation (mm) of deep soil layer

        W,   Total soil moisture (mm)

        WU,  Soil moisture (mm) of upper soil layer

        WL,  Soil moisture (mm) of lower soil layer

        WD,  Soil moisture (mm) of deep soil layer

        R,   Total runoff (mm) of each time step

        RS,  Surface runoff (mm) of each time step

        RI,  Interflow (mm) of each time step

        RG,  Underground runoff (mm) of each time step

        Q,   Total runoff (m^3/s) at the outlet of the basin

        QS,  Surface runoff (m^3/s) at the outlet of the basin

        QI,  Interflow runoff (m^3/s) at the outlet of the basin

        QG,  Underground runoff (m^3/s) at the outlet of the basin
}
\description{
XAJ: An R implementation of three-source Xinanjiang model by Renjun Zhao.

An R implementation of three-source Xinanjiang model
             by Renjun Zhao, used for daily streamflow simulation.
}
\details{
This function is an R implementation of the Xinanjiang (XAJ)
         hydrological model. The lumped XAJ model has 13 parameters, including:

         KC,   Ratio of potential evap to pan evap

         IM,   Fraction of impermeable area

         WUM,  Soil moisture capacity of upper layer

         WLM,  Soil moisture capacity of lower layer

         WDM,  Soil moisture capacity of deep layer

         C,    Coefficient of deep evap

         B,    Exponent of the soil moisture storage capacity curve

         SM,   Areal mean free water capacity of the surface soil layer

         EX,   Exponent of the free water capacity curve

         KI,   outflow coefficients of the free water storage to interflow

         KG,   outflow coefficients of the free water storage to groundwater

         CI,   recession constant of the lower interflow storage

         CG,   recession constant of groundwater storage

         If use the instantaneous unit hydrograph (IUH) of Nash for routing
         of surface runoff, should provided two other parameters:

         N,    number of reservoirs in the instantaneous unit hydrograph

         NK,   common storage coefficient in the instantaneous unit hydrograph

         Else should provided the whole unit hydrograph defined by user,
         and set \code{full.UH} become \code{TRUE}.

         The parameter \code{params} must be a numeric vector looks like:

         \code{c(KC, IM, WUM, WLM, WDM, C, B, SM, EX, KI, KG, CI, CG, N, NK)}

         when use the instantaneous unit hydrograph of Nash, or looks like:

         \code{c(KC, IM, WUM, WLM, WDM, C, B, SM, EX, KI, KG, CI, CG, UH_1, UH_2, ..., UH_n)}

         UH_1, UH_2, ..., UH_n means the series of the unit hydrograph.
}
\references{
Zhao and Liu, 1995. The Xinanjiang model, Computer Models of Watershed Hydrology, Water Resources Publication, Highlands Ranch, CO (1995), pp. 215-232

Zhao and Liu, 1995. The Xinanjiang model, Computer Models of
            Watershed Hydrology, Water Resources Publication, Highlands
            Ranch, CO (1995), pp. 215-232
}
\keyword{Xinanjiang}
\keyword{hydrological}
\keyword{model}

