% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XAJ.R
\name{XAJ_param_range}
\alias{XAJ_param_range}
\title{Get upper and lower boundary of the parameters of XAJ model.}
\usage{
XAJ_param_range()
}
\description{
Get upper and lower boundary of the parameters of XAJ model.
             Meaning of the parameters see `?XAJ`.
}

